+++ estilos/panel/colección.css
