+++ func/util/almacenamiento.js
