+++ func/util/desplazamiento.js
