+++ index.html
@@ -4,10 +4,10 @@
 <head>
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="shortcut icon" href="archivos/imágenes/logo.png">
+	<link rel="shortcut icon" href="archivos/logo.png">
 	<link rel="stylesheet" href="estilos/principal.css">
-	<link rel="stylesheet" href="estilos/colección.css">
-	<link rel="stylesheet" href="estilos/oscuro.css" id="tema">
+	<link rel="stylesheet" href="estilos/panel/colección.css">
+	<link rel="stylesheet" href="estilos/temas/oscuro.css" id="tema">
 	<title> Cronos - Líneas Temporales </title>
 </head>
 
@@ -16,7 +16,7 @@
 
 	<header class="cabecera">
 		<div id="inicio" class="cabecera">
-			<img id="logo" src="archivos/imágenes/logo.png" alt="Logo-Waifu Temporal">
+			<img id="logo" src="archivos/logo.png" alt="Logo-Waifu Temporal">
 			<h1> Cronos </h1>
 		</div>
 		<div>
@@ -24,9 +24,9 @@
 			<button id="creador" class="botón"> Nuevo Tempo </button>
 		</div>
 		<div id="opciones">
-			<img id="botón_idioma" src="archivos/imágenes/banderas/es.png" alt="Idioma">
+			<img id="botón_idioma" src="archivos/banderas/es.png" alt="Idioma">
 			<ul id="lista_idioma" class="oculto"></ul>
-			<input id="botón_tema" type="image" src="archivos/vectores/modo_claro.svg" alt="Cambiar a modo claro">
+			<input id="botón_tema" type="image" src="archivos/temas/modo_claro.svg" alt="Cambiar a modo claro">
 		</div>
 	</header>
 
@@ -44,7 +44,7 @@
 		</div>
 	</main>
 
-	<script src="funciones/principal.js" type="module"></script>
+	<script src="func/principal.js" type="module"></script>
 	<script src="https://kit.fontawesome.com/9fbde9ecdb.js" crossorigin="anonymous"></script>
 </body>
